requirements.txt
PyExecJS==1.1.0
PyYAML==3.11
gevent==1.0.1
greenlet==0.4.5
requests==2.5.3
wsgiref==0.1.2
